# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/loginmobile": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "8eb4b6165db3b40e63a701aeca15a5b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/uploadimage": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Upload Image",
                "operationId": "8eb4b6165db3b40e63a701aeca15a5b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_user": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "id_tps": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "photos": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "photo": {
                                                    "type": "string",
                                                    "format": "base64-encoded-string"
                                                },
                                                "filename": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "photo",
                                                "filename"
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Upload image"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/dapil": {
            "get": {
              "tags": ["Dapil"],
              "summary": "Get Dapil",
              "operationId": "8eb4b6165db3b40e63a701aeca15a5b6",
              "responses": {
                "200": {
                  "description": "Successful retrieval of Dapil data"
                },
                "401": {
                  "description": "Unauthorized"
                }
              }
            }
        },
        "/api/kota": {
            "get": {
              "tags": ["Dapil"],
              "summary": "Get kota",
              "operationId": "8eb4b6165db3b40e63a701aeca15a5b6",
              "responses": {
                "200": {
                  "description": "Successful retrieval of Kota data"
                },
                "401": {
                  "description": "Unauthorized"
                }
              }
            }
        }
    }
}